+++ amethyst_audio/src/systems/mod.rs
@@ -1,6 +1,9 @@
 //! `amethyst` audio ecs systems
 
-pub use self::{audio::AudioSystem, dj::DjSystem};
+pub use self::{
+    audio::{AudioSystem, AudioSystemDesc},
+    dj::{DjSystem, DjSystemDesc},
+};
 
 mod audio;
 mod dj;