+++ amethyst_core/src/frame_limiter.rs
@@ -26,7 +26,8 @@
 //! # struct GameState;
 //! # impl SimpleState for GameState {}
 //! # fn main() -> amethyst::Result<()> {
-//! let mut game = Application::build("./", GameState)?
+//! let assets_dir = "./";
+//! let mut game = Application::build(assets_dir, GameState)?
 //!     .with_frame_limit(
 //!         FrameRateLimitStrategy::SleepAndYield(Duration::from_millis(2)),
 //!         144,