+++ amethyst_rendy/src/bundle.rs
@@ -17,7 +17,7 @@ use crate::{
 };
 use amethyst_assets::Processor;
 use amethyst_core::{
-    ecs::{DispatcherBuilder, Resources},
+    ecs::{DispatcherBuilder, World},
     SystemBundle,
 };
 use amethyst_error::{format_err, Error};
@@ -56,7 +56,11 @@ impl<B: Backend> RenderingBundle<B> {
 }
 
 impl<'a, 'b, B: Backend> SystemBundle<'a, 'b> for RenderingBundle<B> {
-    fn build(mut self, builder: &mut DispatcherBuilder<'a, 'b>) -> Result<(), Error> {
+    fn build(
+        mut self,
+        world: &mut World,
+        builder: &mut DispatcherBuilder<'a, 'b>,
+    ) -> Result<(), Error> {
         builder.add(MeshProcessor::<B>::default(), "mesh_processor", &[]);
         builder.add(TextureProcessor::<B>::default(), "texture_processor", &[]);
         builder.add(Processor::<Material>::new(), "material_processor", &[]);
@@ -70,7 +74,7 @@ impl<'a, 'b, B: Backend> SystemBundle<'a, 'b> for RenderingBundle<B> {
         builder.add_barrier();
 
         for plugin in &mut self.plugins {
-            plugin.on_build(builder)?;
+            plugin.on_build(world, builder)?;
         }
 
         builder.add_thread_local(RenderingSystem::<B, _>::new(self.into_graph_creator()));
@@ -83,22 +87,22 @@ struct RenderingBundleGraphCreator<B: Backend> {
 }
 
 impl<B: Backend> GraphCreator<B> for RenderingBundleGraphCreator<B> {
-    fn rebuild(&mut self, res: &Resources) -> bool {
+    fn rebuild(&mut self, world: &World) -> bool {
         let mut rebuild = false;
         for plugin in self.plugins.iter_mut() {
-            rebuild = plugin.should_rebuild(res) || rebuild;
+            rebuild = plugin.should_rebuild(world) || rebuild;
         }
         rebuild
     }
 
-    fn builder(&mut self, factory: &mut Factory<B>, res: &Resources) -> GraphBuilder<B, Resources> {
+    fn builder(&mut self, factory: &mut Factory<B>, world: &World) -> GraphBuilder<B, World> {
         if self.plugins.is_empty() {
             log::warn!("RenderingBundle is configured to display nothing. Use `with_plugin` to add functionality.");
         }
 
         let mut plan = RenderPlan::new();
         for plugin in self.plugins.iter_mut() {
-            plugin.on_plan(&mut plan, factory, res).unwrap();
+            plugin.on_plan(&mut plan, factory, world).unwrap();
         }
         plan.build(factory).unwrap()
     }
@@ -111,12 +115,16 @@ impl<B: Backend> GraphCreator<B> for RenderingBundleGraphCreator<B> {
 /// and signalling when the graph has to be rebuild.
 pub trait RenderPlugin<B: Backend>: std::fmt::Debug {
     /// Hook for adding systems and bundles to the dispatcher.
-    fn on_build<'a, 'b>(&mut self, _builder: &mut DispatcherBuilder<'a, 'b>) -> Result<(), Error> {
+    fn on_build<'a, 'b>(
+        &mut self,
+        _world: &mut World,
+        _builder: &mut DispatcherBuilder<'a, 'b>,
+    ) -> Result<(), Error> {
         Ok(())
     }
 
     /// Hook for providing triggers to rebuild the render graph.
-    fn should_rebuild(&mut self, _res: &Resources) -> bool {
+    fn should_rebuild(&mut self, _world: &World) -> bool {
         false
     }
 
@@ -125,7 +133,7 @@ pub trait RenderPlugin<B: Backend>: std::fmt::Debug {
         &mut self,
         plan: &mut RenderPlan<B>,
         factory: &mut Factory<B>,
-        res: &Resources,
+        world: &World,
     ) -> Result<(), Error>;
 }
 
@@ -178,7 +186,7 @@ impl<B: Backend> RenderPlan<B> {
         target_plan.add_extension(Box::new(closure));
     }
 
-    fn build(self, factory: &Factory<B>) -> Result<GraphBuilder<B, Resources>, Error> {
+    fn build(self, factory: &Factory<B>) -> Result<GraphBuilder<B, World>, Error> {
         let mut ctx = PlanContext {
             target_metadata: self
                 .targets
@@ -234,7 +242,7 @@ struct PlanContext<B: Backend> {
     passes: HashMap<Target, EvaluationState>,
     outputs: HashMap<TargetImage, ImageId>,
     #[get_mut = "pub"]
-    graph: GraphBuilder<B, Resources>,
+    graph: GraphBuilder<B, World>,
 }
 
 impl<B: Backend> PlanContext<B> {
@@ -260,7 +268,7 @@ impl<B: Backend> PlanContext<B> {
     fn submit_pass(
         &mut self,
         target: Target,
-        pass: RenderPassNodeBuilder<B, Resources>,
+        pass: RenderPassNodeBuilder<B, World>,
     ) -> Result<(), Error> {
         match self.passes.get(&target) {
             None => {}
@@ -426,8 +434,8 @@ impl<'a, B: Backend> TargetPlanContext<'a, B> {
     /// This is useful for adding custom rendering nodes
     /// that are not just standard graphics render passes,
     /// e.g. for compute dispatch.
-    pub fn graph(&mut self) -> &mut GraphBuilder<B, Resources> {
-        self.plan_context.graph_mut()
+    pub fn graph(&mut self) -> &mut GraphBuilder<B, World> {
+        self.plan_context.graph()
     }
 
     /// Retrieve render target metadata, e.g. size.
@@ -644,7 +652,7 @@ impl<B: Backend> TargetPlan<B> {
 #[derive(Debug)]
 pub enum RenderableAction<B: Backend> {
     /// Register single render group for evaluation during target rendering
-    RenderGroup(Box<dyn RenderGroupBuilder<B, Resources>>),
+    RenderGroup(Box<dyn RenderGroupBuilder<B, World>>),
 }
 
 impl<B: Backend> RenderableAction<B> {
@@ -667,7 +675,7 @@ pub trait IntoAction<B: Backend> {
     fn into(self) -> RenderableAction<B>;
 }
 
-impl<B: Backend, G: RenderGroupBuilder<B, Resources> + 'static> IntoAction<B> for G {
+impl<B: Backend, G: RenderGroupBuilder<B, World> + 'static> IntoAction<B> for G {
     fn into(self) -> RenderableAction<B> {
         RenderableAction::RenderGroup(Box::new(self))
     }
@@ -827,7 +835,7 @@ mod tests {
 
         let planned_graph = plan.build(&factory).unwrap();
 
-        let mut manual_graph = GraphBuilder::<DefaultBackend, Resources>::new();
+        let mut manual_graph = GraphBuilder::<DefaultBackend, World>::new();
         let color = manual_graph.create_image(kind, 1, Format::Rgb8Unorm, None);
         let depth = manual_graph.create_image(
             kind,
@@ -902,7 +910,7 @@ mod tests {
 
         let planned_graph = plan.build(&factory).unwrap();
 
-        let mut manual_graph = GraphBuilder::<DefaultBackend, Resources>::new();
+        let mut manual_graph = GraphBuilder::<DefaultBackend, World>::new();
         let depth = manual_graph.create_image(
             window_kind,
             1,