+++ amethyst_rendy/src/pass/flat2d.rs
@@ -11,7 +11,7 @@ use crate::{
 };
 use amethyst_assets::AssetStorage;
 use amethyst_core::{
-    ecs::{Join, Read, ReadExpect, ReadStorage, Resources, SystemData},
+    ecs::{Join, Read, ReadExpect, ReadStorage, SystemData, World},
     transform::Transform,
     Hidden, HiddenPropagate,
 };
@@ -36,19 +36,19 @@ use thread_profiler::profile_scope;
 #[derivative(Default(bound = ""))]
 pub struct DrawFlat2DDesc;
 
-impl<B: Backend> RenderGroupDesc<B, Resources> for DrawFlat2DDesc {
+impl<B: Backend> RenderGroupDesc<B, World> for DrawFlat2DDesc {
     fn build(
         self,
         _ctx: &GraphContext<B>,
         factory: &mut Factory<B>,
         _queue: QueueId,
-        _aux: &Resources,
+        _aux: &World,
         framebuffer_width: u32,
         framebuffer_height: u32,
         subpass: hal::pass::Subpass<'_, B>,
         _buffers: Vec<NodeBuffer>,
         _images: Vec<NodeImage>,
-    ) -> Result<Box<dyn RenderGroup<B, Resources>>, failure::Error> {
+    ) -> Result<Box<dyn RenderGroup<B, World>>, failure::Error> {
         #[cfg(feature = "profiler")]
         profile_scope!("build");
 
@@ -87,14 +87,14 @@ pub struct DrawFlat2D<B: Backend> {
     sprites: OneLevelBatch<TextureId, SpriteArgs>,
 }
 
-impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2D<B> {
+impl<B: Backend> RenderGroup<B, World> for DrawFlat2D<B> {
     fn prepare(
         &mut self,
         factory: &Factory<B>,
         _queue: QueueId,
         index: usize,
         _subpass: hal::pass::Subpass<'_, B>,
-        resources: &Resources,
+        world: &World,
     ) -> PrepareResult {
         #[cfg(feature = "profiler")]
         profile_scope!("prepare opaque");
@@ -117,9 +117,9 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2D<B> {
             ReadStorage<'_, SpriteRender>,
             ReadStorage<'_, Transform>,
             ReadStorage<'_, Tint>,
-        )>::fetch(resources);
+        )>::fetch(world);
 
-        self.env.process(factory, index, resources);
+        self.env.process(factory, index, world);
 
         let sprites_ref = &mut self.sprites;
         let textures_ref = &mut self.textures;
@@ -147,7 +147,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2D<B> {
                     )?;
                     let (tex_id, _) = textures_ref.insert(
                         factory,
-                        resources,
+                        world,
                         texture,
                         hal::image::Layout::ShaderReadOnlyOptimal,
                     )?;
@@ -158,7 +158,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2D<B> {
                 });
         }
 
-        self.textures.maintain(factory, resources);
+        self.textures.maintain(factory, world);
 
         {
             #[cfg(feature = "profiler")]
@@ -181,7 +181,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2D<B> {
         mut encoder: RenderPassEncoder<'_, B>,
         index: usize,
         _subpass: hal::pass::Subpass<'_, B>,
-        _resources: &Resources,
+        _world: &World,
     ) {
         #[cfg(feature = "profiler")]
         profile_scope!("draw opaque");
@@ -200,7 +200,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2D<B> {
         }
     }
 
-    fn dispose(self: Box<Self>, factory: &mut Factory<B>, _aux: &Resources) {
+    fn dispose(self: Box<Self>, factory: &mut Factory<B>, _world: &World) {
         unsafe {
             factory.device().destroy_graphics_pipeline(self.pipeline);
             factory
@@ -215,19 +215,19 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2D<B> {
 #[derivative(Default(bound = ""))]
 pub struct DrawFlat2DTransparentDesc;
 
-impl<B: Backend> RenderGroupDesc<B, Resources> for DrawFlat2DTransparentDesc {
+impl<B: Backend> RenderGroupDesc<B, World> for DrawFlat2DTransparentDesc {
     fn build(
         self,
         _ctx: &GraphContext<B>,
         factory: &mut Factory<B>,
         _queue: QueueId,
-        _aux: &Resources,
+        _world: &World,
         framebuffer_width: u32,
         framebuffer_height: u32,
         subpass: hal::pass::Subpass<'_, B>,
         _buffers: Vec<NodeBuffer>,
         _images: Vec<NodeImage>,
-    ) -> Result<Box<dyn RenderGroup<B, Resources>>, failure::Error> {
+    ) -> Result<Box<dyn RenderGroup<B, World>>, failure::Error> {
         #[cfg(feature = "profiler")]
         profile_scope!("build_trans");
 
@@ -268,14 +268,14 @@ pub struct DrawFlat2DTransparent<B: Backend> {
     change: util::ChangeDetection,
 }
 
-impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2DTransparent<B> {
+impl<B: Backend> RenderGroup<B, World> for DrawFlat2DTransparent<B> {
     fn prepare(
         &mut self,
         factory: &Factory<B>,
         _queue: QueueId,
         index: usize,
         _subpass: hal::pass::Subpass<'_, B>,
-        resources: &Resources,
+        world: &World,
     ) -> PrepareResult {
         #[cfg(feature = "profiler")]
         profile_scope!("prepare transparent");
@@ -288,9 +288,9 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2DTransparent<B> {
                 ReadStorage<'_, SpriteRender>,
                 ReadStorage<'_, Transform>,
                 ReadStorage<'_, Tint>,
-            )>::fetch(resources);
+            )>::fetch(world);
 
-        self.env.process(factory, index, resources);
+        self.env.process(factory, index, world);
         self.sprites.swap_clear();
         let mut changed = false;
 
@@ -316,7 +316,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2DTransparent<B> {
                     )?;
                     let (tex_id, this_changed) = textures_ref.insert(
                         factory,
-                        resources,
+                        world,
                         texture,
                         hal::image::Layout::ShaderReadOnlyOptimal,
                     )?;
@@ -327,7 +327,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2DTransparent<B> {
                     sprites_ref.insert(tex_id, batch_data.drain(..));
                 });
         }
-        self.textures.maintain(factory, resources);
+        self.textures.maintain(factory, world);
         changed = changed || self.sprites.changed();
 
         {
@@ -350,7 +350,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2DTransparent<B> {
         mut encoder: RenderPassEncoder<'_, B>,
         index: usize,
         _subpass: hal::pass::Subpass<'_, B>,
-        _resources: &Resources,
+        _world: &World,
     ) {
         #[cfg(feature = "profiler")]
         profile_scope!("draw transparent");
@@ -369,7 +369,7 @@ impl<B: Backend> RenderGroup<B, Resources> for DrawFlat2DTransparent<B> {
         }
     }
 
-    fn dispose(self: Box<Self>, factory: &mut Factory<B>, _aux: &Resources) {
+    fn dispose(self: Box<Self>, factory: &mut Factory<B>, _aux: &World) {
         unsafe {
             factory.device().destroy_graphics_pipeline(self.pipeline);
             factory