+++ examples/locale/main.rs
@@ -2,7 +2,7 @@
 
 use amethyst::{
     assets::{AssetStorage, Handle, Loader, Processor, ProgressCounter},
-    ecs::{Read, ReadExpect},
+    ecs::{Read, ReadExpect, WorldExt},
     locale::*,
     prelude::*,
     utils::application_root_dir,
@@ -27,7 +27,7 @@ impl Example {
 
 impl SimpleState for Example {
     fn on_start(&mut self, data: StateData<'_, GameData<'_, '_>>) {
-        data.world.add_resource(AssetStorage::<Locale>::new());
+        data.world.insert(AssetStorage::<Locale>::new());
         let mut progress_counter = ProgressCounter::default();
         self.handle_en = Some(data.world.exec(
             |(loader, storage): (ReadExpect<'_, Loader>, Read<'_, AssetStorage<Locale>>)| {
@@ -58,10 +58,20 @@ impl SimpleState for Example {
             let store = data.world.read_resource::<AssetStorage<Locale>>();
             for h in [&self.handle_en, &self.handle_fr].iter() {
                 if let Some(locale) = h.as_ref().and_then(|h| store.get(h)) {
-                    let mut bundle = FluentBundle::new(&[""]);
-                    bundle.add_resource(&locale.resource).unwrap();
-                    println!("{}", bundle.format("hello", None).unwrap().0);
-                    println!("{}", bundle.format("bye", None).unwrap().0);
+                    let bundle = &locale.bundle;
+                    let msg_hello = bundle
+                        .get_message("hello")
+                        .expect("Failed to load message for hello");
+                    let msg_bye = bundle
+                        .get_message("bye")
+                        .expect("Failed to load message for bye");
+                    let hello_value = msg_hello.value.expect("Hello message has no value");
+                    let bye_value = msg_bye.value.expect("Bye message has no value");
+
+                    let mut errors = vec![];
+                    println!("{}", bundle.format_pattern(hello_value, None, &mut errors));
+                    println!("{}", bundle.format_pattern(bye_value, None, &mut errors));
+                    assert_eq!(errors.len(), 0);
                 }
             }
             Trans::Quit
@@ -74,11 +84,11 @@ impl SimpleState for Example {
 fn main() -> Result<(), Error> {
     amethyst::start_logger(Default::default());
 
-    let assets_directory = application_root_dir()?.join("examples/assets");
+    let assets_dir = application_root_dir()?.join("examples/assets");
 
     let game_data = GameDataBuilder::default().with(Processor::<Locale>::new(), "proc", &[]);
 
-    let mut game = Application::new(assets_directory, Example::new(), game_data)?;
+    let mut game = Application::new(assets_dir, Example::new(), game_data)?;
     game.run();
     Ok(())
 }