+++ docs/CHANGELOG.md
@@ -12,7 +12,11 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 
 ### Major breaking changes
 
+* Systems needing initialization with world resources must go through a `SystemDesc` intermediate builder. ([#1780])
+
 ### Added
+
+* `SystemDesc` proc macro derive to simplify defining `SystemDesc`s. ([#1780])
 
 ### Changed
 
@@ -21,6 +25,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 ### Fixed
 
 [#1848]: https://github.com/amethyst/amethyst/pull/1848
+[#1780]: https://github.com/amethyst/amethyst/pull/1780
 
 ## [0.12.0] - 2019-07-30